# openapi.yaml
openapi: 3.0.3
info:
  title: The Vanity E-commerce API
  description: |
    RESTful API for The Vanity E-commerce Platform.
    This API provides endpoints for managing products, users, orders, and admin operations.
  version: 1.0.0
  contact:
    name: The Vanity Support
    email: support@thevanity.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.thevanity.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
  schemas:
...
    Category:
      type: object
      required:
        - name
      properties:
        _id:
          type: string
        name:
          type: string
          example: "Makeup"
        slug:
          type: string
          example: "makeup"
        description:
          type: string
        parentId:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
        sortOrder:
          type: integer
          default: 0
        image:
          type: string
          format: uri
        attributes:
          type: array
          items:
            type: string
        productCount:
          type: integer
          readOnly: true
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        statusCode:
          type: integer
          example: 400
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
            email:
              type: string
            role:
              type: string
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
                phone:
                  type: string
                  example: "+91 98765 43210"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== USER PROFILE ====================
...
  /cart:
    get:
      tags:
        - Cart
      summary: Get user cart
      description: Retrieve authenticated user's cart
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object

    post:
      tags:
        - Cart
      summary: Add to cart
      description: Add item to user's cart
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - quantity
              properties:
                productId:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Item added to cart
  /cart/{itemId}:
    put:
      tags:
        - Cart
      summary: Update cart item
      description: Update quantity of cart item
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Cart item updated

    delete:
      tags:
        - Cart
      summary: Remove from cart
      description: Remove item from cart
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item removed from cart

  # ==================== WISHLIST ====================
...
  /admin/categories:
    get:
      tags:
        - Admin - Categories
      summary: Get all categories (Admin)
      description: Retrieve all categories with admin details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      tags:
        - Admin - Categories
      summary: Create category
      description: Create a new category (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '201':
          description: Category created successfully
  /admin/categories/{id}:
    put:
      tags:
        - Admin - Categories
      summary: Update category
      description: Update existing category (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Category updated successfully

    delete:
      tags:
        - Admin - Categories
      summary: Delete category
      description: Delete a category (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully

  # ==================== ADMIN - ORDERS ====================
...